requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116

torch==1.13.1+cu116
torchvision==0.14.1+cu116
timm==0.4.9
numpy==1.23.4
scipy==1.10.1
imgaug==0.4.0
pillow==9.3.0
matplotlib==3.6.3
hub==3.0.1
pandas==1.5.2
six==1.16.0
wandb
tqdm